add export download btn
- Abstract download btn code to its own component
- Add to import/export dialog
- Fix poracle return when exporting multiple features
